+++ reqwest-retry/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- `RetryableStrategy` which allows for custom retry decisions based on the response that a request got
 
 ## [0.2.1] - 2022-12-01
 
@@ -12,7 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Classify `io::Error`s and `hyper::Error(Canceled)` as transient
 
 ## [0.2.0] - 2022-11-15
-
 ### Changed
 - Updated `reqwest-middleware` to `0.2.0`
