+++ reqwest-retry/src/lib.rs
@@ -27,8 +27,13 @@
 
 mod middleware;
 mod retryable;
+mod retryable_strategy;
 
 pub use retry_policies::{policies, RetryPolicy};
 
 pub use middleware::RetryTransientMiddleware;
 pub use retryable::Retryable;
+pub use retryable_strategy::{
+    default_on_request_failure, default_on_request_success, DefaultRetryableStrategy,
+    RetryableStrategy,
+};
